+++ components/SystemSettingStorageInterface.php
@@ -19,7 +19,11 @@ interface SystemSettingStorageInterface
     
     const TYPE_ARRAY_NAME = 'array';
     
+    public function edit(string $key, int $type, ?string $value, ?string $default, ?string $description);
+    
     public function get($key);
     
     public function set($key, $value);
+    
+    public function remove($key);
 }