+++ tests/_bootstrap.php
@@ -1,0 +1,11 @@
+<?php
+
+// defined('YII_DEBUG') or define('YII_DEBUG', true);
+// defined('YII_ENV') or define('YII_ENV', 'test');
+
+// require(__DIR__ . '/../vendor/autoload.php');
+// require(__DIR__ . '/../vendor/yiisoft/yii2/Yii.php');
+
+// var_dump(__DIR__ . '/../../../vendor/autoload.php'); die();
+require(__DIR__ . '/../../../../../vendor/autoload.php');
+// require(__DIR__ . '/../../../../../vendor/yiisoft/yii2/Yii.php');