+++ views/setting-storage/update.php
@@ -6,6 +6,7 @@ use yii\helpers\Html;
 /* @var $model sergeylisitsyn\settingsStorage\models\SettingStorage */
 
 $this->title = 'Update Setting Storage: ' . $model->name;
+$this->params['breadcrumbs'][] = ['label' => 'Setting Storage module', 'url' => ['/admin/settings-storage']];
 $this->params['breadcrumbs'][] = ['label' => 'Setting Storages', 'url' => ['index']];
 $this->params['breadcrumbs'][] = ['label' => $model->name, 'url' => ['view', 'id' => $model->id]];
 $this->params['breadcrumbs'][] = 'Update';